bodovi implementirani

+++ app/src/main/java/ba/etf/rma21/projekat/data/models/OdgovorKviz.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 class OdgovorKviz(
     @SerializedName("odgovor") val odgovor: Int,
     @SerializedName("pitanje") val pitanjeId: Int,
-    @SerializedName("bodovi") val osvojeniBodovi: Float
+    @SerializedName("bodovi") val osvojeniBodovi: Int
 ){}
